split form components out further, move elements from filesandruns state into addpredictorsformstate.
