+++ neuroscout/frontend/package.json
@@ -36,12 +36,14 @@
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
     "react-scripts-ts": "3.1.0",
+    "source-map-explorer": "^2.0.1",
     "tslint-etc": "^1.5.6",
     "tsutils": "^2.26.2",
     "typescript": "^3"
   },
   "scripts": {
     "start": "react-scripts-ts start",
+    "analyze": "source-map-explorer 'build/static/js/*.js'",
     "build": "react-scripts-ts build",
     "test": "react-scripts-ts test --env=jsdom",
     "eject": "react-scripts-ts eject"
