+++ neuroscout/frontend/src/predictor_collection/PredictorDescriptionForm.tsx
@@ -10,7 +10,7 @@ type PDFProps = {
 export class PredictorDescriptionForm extends React.Component<PDFProps, {} > {
   render() {
     let formList = this.props.predictors.map((predictor, index) => (
-        <Form.Item label={`Description for predictor ${predictor}`}>
+        <Form.Item key={`PDF${index}`} label={`Description for predictor ${predictor}`}>
           <Input
             value={this.props.descriptions[index]}
             onChange={(e) => { this.props.updateDescription(index, e.target.value); }}
