+++ neuroscout/frontend/tslint.json
@@ -52,7 +52,7 @@
         "no-switch-case-fall-through": true,
         "no-trailing-whitespace": false,
         "no-unused-expression": false,
-        "no-unused-declaration": true,
+        "no-unused-declaration": false,
         "no-use-before-declare": true,
         "one-line": [
             true,
